+++ src/com/sap/piper/DefaultValueCache.groovy
@@ -6,6 +6,8 @@ import com.sap.piper.MapUtils
 class DefaultValueCache implements Serializable {
     private static DefaultValueCache instance
 
+    //static CommonPipelineEnvironment commonPipelineEnvironment = new CommonPipelineEnvironment()
+
     private Map defaultValues
 
     private DefaultValueCache(Map defaultValues){
